pyproject.toml
[project]
name = "car-mileage-log-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg>=3.2.9",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "records>=0.6.0",
    "sqlalchemy>=2.0.42",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "ipython>=9.4.0",
    "rich>=14.1.0",
    "sqlacodegen>=3.0.0",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
